# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DAO Governance System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProposalStatus =
  (typeof ProposalStatus)[keyof typeof ProposalStatus];

export const ProposalStatus = {
  active: "active",
  pending: "pending",
  executed: "executed",
  defeated: "defeated",
  queued: "queued",
} as const;

export interface Proposal {
  id: number;
  title: string;
  description: string;
  status: ProposalStatus;
  proposer: string;
  /** @nullable */
  proposerEns?: string | null;
  votesFor: number;
  votesAgainst: number;
  votesAbstain: number;
  quorumRequired: number;
  quorumReached: boolean;
  startTime: string;
  endTime: string;
  createdAt: string;
  /** @nullable */
  executedAt?: string | null;
  actions?: string[];
  /** @nullable */
  timelockDelay?: number | null;
  /** @nullable */
  discussionUrl?: string | null;
  /** @nullable */
  snapshotBlock?: number | null;
}

export interface ProposalInput {
  /** @minLength 5 */
  title: string;
  /** @minLength 20 */
  description: string;
  actions?: string[];
  /** Duration in hours */
  votingDuration: number;
  discussionUrl?: string;
  proposer?: string;
}

export type VoteSupport = (typeof VoteSupport)[keyof typeof VoteSupport];

export const VoteSupport = {
  for: "for",
  against: "against",
  abstain: "abstain",
} as const;

export interface Vote {
  id: number;
  proposalId: number;
  voter: string;
  /** @nullable */
  voterEns?: string | null;
  support: VoteSupport;
  weight: number;
  /** @nullable */
  reason?: string | null;
  createdAt: string;
}

export type VoteInputSupport =
  (typeof VoteInputSupport)[keyof typeof VoteInputSupport];

export const VoteInputSupport = {
  for: "for",
  against: "against",
  abstain: "abstain",
} as const;

export interface VoteInput {
  voter: string;
  support: VoteInputSupport;
  weight: number;
  reason?: string;
}

export interface Comment {
  id: number;
  proposalId: number;
  author: string;
  /** @nullable */
  authorEns?: string | null;
  content: string;
  createdAt: string;
}

export interface CommentInput {
  author: string;
  /** @minLength 1 */
  content: string;
}

export interface Member {
  id: number;
  address: string;
  /** @nullable */
  ens?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  votingPower: number;
  tokenBalance: number;
  proposalsCreated: number;
  votesCast: number;
  /** @nullable */
  delegatedTo: string | null;
  /** @nullable */
  delegatedToEns?: string | null;
  delegatorsCount?: number;
  /** @nullable */
  bio?: string | null;
  createdAt: string;
}

export interface DelegateInput {
  delegateTo: string;
}

export interface TreasuryAllocation {
  asset: string;
  amount: number;
  valueUsd: number;
  percent: number;
}

export interface Treasury {
  totalValueUsd: number;
  ethBalance: number;
  tokenBalance: number;
  stablecoinBalance?: number;
  changePercent24h?: number;
  allocations: TreasuryAllocation[];
}

export type TreasuryTransactionType =
  (typeof TreasuryTransactionType)[keyof typeof TreasuryTransactionType];

export const TreasuryTransactionType = {
  inflow: "inflow",
  outflow: "outflow",
  swap: "swap",
} as const;

export interface TreasuryTransaction {
  id: number;
  type: TreasuryTransactionType;
  asset: string;
  amount: number;
  valueUsd: number;
  address: string;
  txHash: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  proposal_created: "proposal_created",
  vote_cast: "vote_cast",
  proposal_executed: "proposal_executed",
  proposal_defeated: "proposal_defeated",
  delegation_changed: "delegation_changed",
  comment_added: "comment_added",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  actor: string;
  /** @nullable */
  actorEns?: string | null;
  description: string;
  /** @nullable */
  proposalId?: number | null;
  /** @nullable */
  proposalTitle?: string | null;
  createdAt: string;
}

export interface AnalyticsSummary {
  totalProposals: number;
  activeProposals: number;
  totalMembers: number;
  totalVotesCast: number;
  averageParticipation: number;
  successRate: number;
  totalTokensCirculating: number;
  treasuryValueUsd: number;
}

export interface VotingHistoryPoint {
  month: string;
  participation: number;
  proposalsCreated: number;
  proposalsPassed: number;
}

export interface TokenDistributionItem {
  label: string;
  percent: number;
  amount: number;
}

export type ListProposalsParams = {
  status?: ListProposalsStatus;
  search?: string;
  sort?: ListProposalsSort;
};

export type ListProposalsStatus =
  (typeof ListProposalsStatus)[keyof typeof ListProposalsStatus];

export const ListProposalsStatus = {
  active: "active",
  pending: "pending",
  executed: "executed",
  defeated: "defeated",
  all: "all",
} as const;

export type ListProposalsSort =
  (typeof ListProposalsSort)[keyof typeof ListProposalsSort];

export const ListProposalsSort = {
  newest: "newest",
  oldest: "oldest",
  most_votes: "most_votes",
  ending_soon: "ending_soon",
} as const;

export type ListMembersParams = {
  search?: string;
  sort?: ListMembersSort;
};

export type ListMembersSort =
  (typeof ListMembersSort)[keyof typeof ListMembersSort];

export const ListMembersSort = {
  voting_power: "voting_power",
  proposals_created: "proposals_created",
  votes_cast: "votes_cast",
  newest: "newest",
} as const;

export type ListActivityParams = {
  limit?: number;
};