# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DAO Governance System API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemType } from "./activityItemType";

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  actor: string;
  /** @nullable */
  actorEns?: string | null;
  description: string;
  /** @nullable */
  proposalId?: number | null;
  /** @nullable */
  proposalTitle?: string | null;
  createdAt: string;
}