# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DAO Governance System API
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  proposalId: number;
  author: string;
  /** @nullable */
  authorEns?: string | null;
  content: string;
  createdAt: string;
}