# lib/api-zod/src/generated/types/proposalInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DAO Governance System API
 * OpenAPI spec version: 0.1.0
 */

export interface ProposalInput {
  /** @minLength 5 */
  title: string;
  /** @minLength 20 */
  description: string;
  actions?: string[];
  /** Duration in hours */
  votingDuration: number;
  discussionUrl?: string;
  proposer?: string;
}