# lib/api-zod/src/generated/types/votingHistoryPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DAO Governance System API
 * OpenAPI spec version: 0.1.0
 */

export interface VotingHistoryPoint {
  month: string;
  participation: number;
  proposalsCreated: number;
  proposalsPassed: number;
}